Update session via callbacks from mafia output
No longer dependent on having hopefully run a custom mafia script since
the last time session changed. There is a glaring problem with this in
that it means the session will lag one tick behind or so, which I'm
definitely going to need to look in to a solution to.