Stop using exec completely
No longer dependent on other plugins. I was hoping exec was the cause of
a crash I've been dealing with, but apparently not. But this is a big
improvement either way, since you can now check if mafia is running.
Also doesn't require me to expose an inner working function as a
/command, which is nice. And it's probably much more efficient too.